requirements.txt
# PDF and DOCX processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Machine Learning and Transformers
torch>=2.0.0
transformers>=4.30.0

# Logging and utilities
pathlib>=1.0.1 